pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "matplotlib>=3.10.7",
    "numpy>=1.24.0",
    "pandas>=2.3.3",
    "pillow>=10.0.0",
    "scikit-learn>=1.7.2",
    "seaborn>=0.13.2",
    "streamlit>=1.30.0",
    "tensorflow>=2.15.0",
]
